requirements.txt
streamlit
toml
pandas
Authlib
opencv-python
PyYAML
dotenv
colorama